refactor(camera): decompose manager.go (2109 LOC) into 6 cohesive sub-files + add relay smoke test
